# New high level interface for cuDNN