… lib/cudnn/README.md
… lib/cudnn/compat.jl
… lib/cudnn/conv.jl
… lib/cudnn/descriptors.jl
… lib/cudnn/filter.jl
… lib/cudnn/inplace.jl
… lib/cudnn/libcudnn.jl
… lib/cudnn/multiheadattn.jl
… lib/cudnn/nnlib.jl
… lib/cudnn/normalization.jl
… lib/cudnn/optensor.jl
… lib/cudnn/pooling.jl
… lib/cudnn/reduce.jl
… lib/cudnn/rnn.jl
… lib/cudnn/softmax.jl
… lib/cudnn/tensor.jl
… lib/cudnn/util.jl
… test/cudnn.jl
… test/cudnn/activation.jl
… test/cudnn/convolution.jl
… test/cudnn/dropout.jl
… test/cudnn/inplace.jl
… test/cudnn/multiheadattn.jl
… test/cudnn/nnlib.jl
… test/cudnn/normalization.jl
… test/cudnn/optensor.jl
… test/cudnn/pooling.jl
… test/cudnn/reduce.jl
… test/cudnn/rnn.jl
… test/cudnn/softmax.jl