Please see [README.md](https://github.com/denizyuret/CUDA.jl/blob/dy/cudnn/lib/cudnn/README.md) for some design decisions, comments welcome. I tried not to break any existing code by defining deprecated methods. Here is a todo list.
- [x] README: Describes design choices
- [x] util.jl: Completed `cudnnDataType()`. Replace `c_null(x),cu_null(x)->something(x, C/CU_NULL)` , `cudnnWorkspace->@workspace/cudnnTempSpace`, `scalr -> scalingParameter`, `@retry -> @retry_reclaim`, `DevArray->CuArray`.
- [x] base.jl
- [x] descriptors.jl: defines all descriptors consistently using `@cudnnDescriptor` macro.
- [x] ~~tensor.jl (old)~~: defines TensorDesc, add, op, reduce wrappers. These are defined in separate files in my implementation.
- [x] ~~filter.jl~~: merged with tensor.jl
- [x] tensor.jl: alternative array based constructors for Tensor and Filter. Define compat function with old TensorDesc constructor. **tested**
- [x] inplace.jl: set, scale, add. **tested**
- [x] optensor.jl: **tested**
- [x] reduce.jl: **tested**
- [x] activation.jl: **tested**
- [x] softmax.jl: **tested** 
- [x] pooling.jl: removed NNlib dependency. **tested**
- [x] dropout.jl: **tested**
- [x] ~~conv.jl~~ -> convolution.jl: including conv-bias-act: removed NNlib dependency. **tested**
- [x] rnn.jl: new interface done. old interface complicated and incomplete, I am skipping deprecated methods for this one. **tested**
- [x] multiheadattn.jl: **tested**
- [x] ~~batchnorm.jl~~ -> normalization.jl: **tested**  ~~TODO: rethink arg order: adopted one from tf.nn~~ batchnorm needs to be deprecated and implemented in terms of normalization.jl: leaving this to nnlib.
- [x] nnlib.jl: **tested**. should rename test file test/cudnn.jl -> test/cudnn/nnlib.jl. ~~TODO: profile and pick faster version of softmax.~~
- [x] ~~compat.jl~~ -> **tested** merged into nnlib.jl
- [x] CUDNN.jl
- [x] docs
- [x] tests: julia17 --project -e 'using Pkg; Pkg.API.test(; test_args=\`--memcheck --jobs=1 cudnn\`)                                          
- [x] deprecations
- [x] profiling