@denizyuret 2020-11-07
@maleadt I need some advice on https://github.com/denizyuret/Knet.jl/blob/2dec7d58a9d40899c5c38d6f4f17af463163409f/src/cudnn/dropout.jl#L66-L88

As far as I can understand from cudnn docs, `cudnnDropoutState` should be similar to `Random.GLOBAL_RNG`. There is this warning about concurrently running dropouts using the same state. I don't know if this concerns threads on the CPU or streams on the GPU or both. How should we handle this? Did you have a similar issue with `CURAND`?