@denizyuret 2020-11-10
> If you rebase this on `master`, CI will work.

I occasionally merge master. I don't know how/why to rebase.