@denizyuret 2020-11-14
> > I occasionally merge master. I don't know how/why to rebase.
> 
> `git rebase master` (after pulling it, else `git fetch` && `git rebase origin/master`). Better do so, to avoid ugly merge commits.

I was able to do a successful `git rebase master` once but now I get scary conflict messages when I try. We'll clean up the history when it is time to review.