@DhairyaLGandhi 2020-12-23
I've been playing with the PR for a couple days now, and I was noticing some performance regressions training with several models, but resnet shows it fairly simply

Current:
```
BenchmarkTools.Trial:
  memory estimate:  676.37 KiB
  allocs estimate:  22051
  --------------
  minimum time:     413.535 ms (0.99% GC)
  median time:      496.596 ms (0.82% GC)
  mean time:        571.457 ms (0.78% GC)
  maximum time:     952.702 ms (0.63% GC)
  --------------
  samples:          5
  evals/sample:     1
```

This:
```
BenchmarkTools.Trial:
  memory estimate:  348.56 KiB
  allocs estimate:  9584
  --------------
  minimum time:     619.104 ms (0.00% GC)
  median time:      685.281 ms (0.00% GC)
  mean time:        868.306 ms (0.00% GC)
  maximum time:     1.484 s (0.00% GC)
  --------------
  samples:          4
  evals/sample:     1
``` 

This is with an input of `(224,224,3,128)`. The problem seems to rise with batch sizes, although in some cudnn bound cases, I see smaller batch outperform the current 🎉