@denizyuret 2020-12-23
I agree Dhairya, however I am not sure about the specifics. There are at
least two types of optimizations in cudnn: one involving the use of tensor
cores (as specified by CUDNN.math_mode) and the other specifying whether
`(x - maximum(x; dims))`, improving numerical stability, is to be used in
the softmax calculation (CUDNN_SOFTMAX_ACCURATE uses this,
CUDNN_SOFTMAX_FAST doesn't).

CUDA.jl specifies three math modes: PEDANTIC, DEFAULT, FAST. The issue is
which softmax algo maps to which math mode. I think we can all agree that
PEDANTIC should map to CUDNN_SOFTMAX_ACCURATE and FAST_MATH should map to
CUDNN_SOFTMAX_FAST. I am not sure about the DEFAULT_MATH. The comment for
default says: "use at least the required precision, and allow reordering
operations for performance". I think by dropping `(x - maximum(x; dims))`
sacrificing numerical stability but not decreasing the precision directly,
CUDNN_SOFTMAX_FAST could map to DEFAULT_MATH. Unless I hear otherwise, that
is what I am going to implement.

On Mon, Dec 21, 2020 at 2:25 PM Dhairya Gandhi <notifications@github.com>
wrote:

> *@DhairyaLGandhi* commented on this pull request.
> ------------------------------
>
> In lib/cudnn/nnlib.jl
> <https://github.com/JuliaGPU/CUDA.jl/pull/523#discussion_r546652984>:
>
> >
>  function softmax!(y::DenseCuArray{T}, x::DenseCuArray{T}; dims=1) where T<:CUDNNFloat
> -  cudnnSoftmaxForward(reshape4D(x), reshape4D(y),
> -                      algo=CUDNN_SOFTMAX_FAST, mode=cudnnSoftmaxMode_t(dims-1))
> -  return y
> +    mode, xsize = softmaxhelper(size(x), dims)
> +    cudnnSoftmaxForward!(reshape(y,xsize), reshape(x,xsize); algo=CUDNN_SOFTMAX_FAST, mode)
>
> I think we should have the current math mode as the default and have
> libraries request the fast paths where possible.
>
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub
> <https://github.com/JuliaGPU/CUDA.jl/pull/523#discussion_r546652984>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAN43JUQIP57E2VQDRKV7YLSV4WAZANCNFSM4TMSJE4Q>
> .
>