@DhairyaLGandhi 2020-12-23
Aside: Have you also noticed the operation fusing work they have introduced in cudnn 8?