@DhairyaLGandhi 2020-12-23
I think by default it should map to the accurate version only because that has been the convention. I doubt there are many cases for ML where the comparative drop in stability would hurt convergence directly, but would you imagine it could improve performance enough to justify it? Specifically I am thinking if we can have some sort of opt-in behaviour downstream so its more in-line with the rest of the machinery.