@denizyuret 2020-12-23
> Aside: Have you also noticed the operation fusing work they have introduced in cudnn 8?

Yes, that seems awesome. Unfortunately I did not have enough time to decipher the documentation to figure out how it works. Do you have any recommended reading?