@denizyuret 2020-12-23
> I think by default it should map to the accurate version only because that has been the convention. I doubt there are many cases for ML where the comparative drop in stability would hurt convergence directly, but would you imagine it could improve performance enough to justify it? Specifically I am thinking if we can have some sort of opt-in behaviour downstream so its more in-line with the rest of the machinery.

As Jeff said [here](https://github.com/JuliaGPU/CUDA.jl/pull/321#issuecomment-669297684) -- there is no convention, we should do what is best for the application (in this case deep learning). I will look at the relative performance vs relative degradation of accuracy (which should be negligible in modern models where the input is normalized) and get back on this.