@DhairyaLGandhi 2020-12-23
I have added the script in this [gist](https://gist.github.com/DhairyaLGandhi/12d06b9389e13130cf4c03c67885be91). Left the Manifest out since it depends on local paths, but basically checking out the PR locally with the latest version of LLVM, GPUCompiler, as well as Zygote and Metalhead