@DhairyaLGandhi 2020-12-23
> we should do what is best for the application

Agree, I am assuming there wouldn't be much appreciable dip in which case the fast defaults should do us good.